+++ NEWS.md
@@ -1,5 +1,5 @@
 # anvl (development version)
-
+  
 ## Breaking changes
 
 * The type system of {anvl} was changed to avoid the problems reported in issue #373.
@@ -7,6 +7,11 @@
   With it, also the promotion behavior of various primitives and API
   functions was improved.
 * `jit_eval()` was removed as it is no longer needed.
+
+## Features
+
+* New functions for the uniform distribution: `nv_dunif()`, `nv_punif()`,
+  and `nv_qunif()`.
 
 ## Tests